tests: test_framework layout cleanup
This splits the huge utils.py into one file per binary. We keep utils.py
for the bulk, and, well utils.

As we added new utilities along the way and two new binaries are yet to
come it really started to be bloated.
