+++ tests/test_framework/utils.py
@@ -4,26 +4,17 @@ Rusty Russell or Christian Decker who wrote most of this (I'd put some sats on
 cdecker), so credits to them ! (MIT licensed)
 """
 import bip32
-import bitcoin
 import coincurve
 import itertools
 import json
 import logging
-import pathlib
-import psycopg2
 import os
-import random
 import re
-import select
 import socket
 import subprocess
 import threading
 import time
 
-from bitcoin.rpc import RawProxy as BitcoinProxy
-from decimal import Decimal
-from ephemeral_port_reserve import reserve
-from nacl.public import PrivateKey as Curve25519Private
 from test_framework import serializations
 from typing import Optional
 
@@ -442,728 +433,3 @@ class TailableProc(object):
         Convenience wrapper for the common case of only seeking a single entry.
         """
         return self.wait_for_logs([regex], timeout)
-
-
-class SimpleBitcoinProxy:
-    """Wrapper for BitcoinProxy to reconnect.
-
-    Long wait times between calls to the Bitcoin RPC could result in
-    `bitcoind` closing the connection, so here we just create
-    throwaway connections. This is easier than to reach into the RPC
-    library to close, reopen and reauth upon failure.
-    """
-
-    def __init__(self, bitcoind_dir, bitcoind_port, *args, **kwargs):
-        self.__btc_conf_file__ = os.path.join(bitcoind_dir, "bitcoin.conf")
-        self.__cookie_path = os.path.join(bitcoind_dir, "regtest", ".cookie")
-        self.__port = bitcoind_port
-        # The internal bitcoind wallet, used to generate blocks and distribute
-        # coins
-        self.wallet_name = "revaultd-tests"
-
-    def __getattr__(self, name):
-        if name.startswith("__") and name.endswith("__"):
-            # Python internal stuff
-            raise AttributeError
-
-        # We want to hit the per-wallet API and python-bitcoinlib will not read
-        # the cookie if we specify a custom URL..
-        with open(self.__cookie_path) as fd:
-            authpair = fd.read()
-        service_url = (
-            f"http://{authpair}@localhost:{self.__port}/wallet" f"/{self.wallet_name}"
-        )
-
-        # Create a callable to do the actual call
-        proxy = BitcoinProxy(
-            btc_conf_file=self.__btc_conf_file__, service_url=service_url
-        )
-
-        def f(*args):
-            return proxy._call(name, *args)
-
-        # Make debuggers show <function bitcoin.rpc.name> rather than <function
-        # bitcoin.rpc.<lambda>>
-        f.__name__ = name
-        return f
-
-
-class BitcoinD(TailableProc):
-    def __init__(self, bitcoin_dir, rpcport=None):
-        TailableProc.__init__(self, bitcoin_dir, verbose=False)
-
-        if rpcport is None:
-            rpcport = reserve()
-
-        self.bitcoin_dir = bitcoin_dir
-        self.rpcport = rpcport
-        self.p2pport = reserve()
-        self.prefix = "bitcoind"
-
-        regtestdir = os.path.join(bitcoin_dir, "regtest")
-        if not os.path.exists(regtestdir):
-            os.makedirs(regtestdir)
-
-        self.cmd_line = [
-            "bitcoind",
-            "-datadir={}".format(bitcoin_dir),
-            "-printtoconsole",
-            "-server",
-            "-logtimestamps",
-            "-rpcthreads=4",
-        ]
-        bitcoind_conf = {
-            "port": self.p2pport,
-            "rpcport": rpcport,
-            "debug": 1,
-            "fallbackfee": Decimal(1000) / bitcoin.core.COIN,
-        }
-        self.conf_file = os.path.join(bitcoin_dir, "bitcoin.conf")
-        with open(self.conf_file, "w") as f:
-            f.write(f"chain=regtest\n")
-            f.write(f"[regtest]\n")
-            for k, v in bitcoind_conf.items():
-                f.write(f"{k}={v}\n")
-
-        self.rpc = SimpleBitcoinProxy(
-            bitcoind_dir=self.bitcoin_dir, bitcoind_port=self.rpcport
-        )
-        self.proxies = []
-
-    def start(self):
-        TailableProc.start(self)
-        self.wait_for_log("Done loading", timeout=TIMEOUT)
-
-        logging.info("BitcoinD started")
-
-    def stop(self):
-        for p in self.proxies:
-            p.stop()
-        self.rpc.stop()
-        return TailableProc.stop(self)
-
-    # wait_for_mempool can be used to wait for the mempool before generating
-    # blocks:
-    # True := wait for at least 1 transation
-    # int > 0 := wait for at least N transactions
-    # 'tx_id' := wait for one transaction id given as a string
-    # ['tx_id1', 'tx_id2'] := wait until all of the specified transaction IDs
-    def generate_block(self, numblocks=1, wait_for_mempool=0):
-        if wait_for_mempool:
-            if isinstance(wait_for_mempool, str):
-                wait_for_mempool = [wait_for_mempool]
-            if isinstance(wait_for_mempool, list):
-                wait_for(
-                    lambda: all(
-                        txid in self.rpc.getrawmempool() for txid in wait_for_mempool
-                    )
-                )
-            else:
-                wait_for(lambda: len(self.rpc.getrawmempool()) >= wait_for_mempool)
-
-        addr = self.rpc.getnewaddress()
-        return self.rpc.generatetoaddress(numblocks, addr)
-
-    def get_coins(self, amount_btc):
-        # subsidy halving is every 150 blocks on regtest, it's a rough estimate
-        # to avoid looping in most cases
-        numblocks = amount_btc // 25 + 1
-        while self.rpc.getbalance() < amount_btc:
-            self.generate_block(numblocks)
-
-    def generate_blocks_censor(self, n, txids):
-        """Generate {n} blocks ignoring {txids}"""
-        fee_delta = 1000000
-        for txid in txids:
-            self.rpc.prioritisetransaction(txid, None, -fee_delta)
-        self.generate_block(n)
-        for txid in txids:
-            self.rpc.prioritisetransaction(txid, None, fee_delta)
-
-    def simple_reorg(self, height, shift=0):
-        """
-        Reorganize chain by creating a fork at height={height} and:
-            - If shift >=0:
-                - re-mine all mempool transactions into {height} + shift
-                  (with shift floored at 1)
-            - Else:
-                - don't re-mine the mempool transactions
-
-        Note that tx's that become invalid at {height} (because coin maturity,
-        locktime etc.) are removed from mempool. The length of the new chain
-        will be original + 1 OR original + {shift}, whichever is larger.
-
-        For example: to push tx's backward from height h1 to h2 < h1,
-        use {height}=h2.
-
-        Or to change the txindex of tx's at height h1:
-        1. A block at height h2 < h1 should contain a non-coinbase tx that can
-            be pulled forward to h1.
-        2. Set {height}=h2 and {shift}= h1-h2
-        """
-        orig_len = self.rpc.getblockcount()
-        old_hash = self.rpc.getblockhash(height)
-        if height + shift > orig_len:
-            final_len = height + shift
-        else:
-            final_len = 1 + orig_len
-
-        self.rpc.invalidateblock(old_hash)
-        self.wait_for_log(
-            r"InvalidChainFound: invalid block=.*  height={}".format(height)
-        )
-        memp = self.rpc.getrawmempool()
-
-        if shift < 0:
-            self.generate_blocks_censor(1 + final_len - height, memp)
-        elif shift == 0:
-            self.generate_block(1 + final_len - height, memp)
-        else:
-            self.generate_blocks_censor(shift, memp)
-            self.generate_block(1 + final_len - (height + shift), memp)
-        self.wait_for_log(r"UpdateTip: new best=.* height={}".format(final_len))
-
-    def startup(self):
-        try:
-            self.start()
-        except Exception:
-            self.stop()
-            raise
-
-        info = self.rpc.getnetworkinfo()
-
-        if info["version"] < 210000:
-            self.rpc.stop()
-            raise ValueError(
-                "bitcoind is too old. At least version 21000"
-                " (v0.21.0) is needed, current version is {}".format(info["version"])
-            )
-
-    def cleanup(self):
-        try:
-            self.stop()
-        except Exception:
-            self.proc.kill()
-        self.proc.wait()
-
-
-class Revaultd(TailableProc):
-    def __init__(
-        self,
-        datadir,
-        stks,
-        cosigs,
-        mans,
-        csv,
-        noise_priv,
-        coordinator_noise_key,
-        coordinator_port,
-        bitcoind,
-        stk_config=None,
-        man_config=None,
-    ):
-        assert stk_config is not None or man_config is not None
-        assert len(stks) == len(cosigs)
-        TailableProc.__init__(self, datadir, verbose=VERBOSE)
-
-        self.prefix = os.path.split(datadir)[-1]
-
-        # The data is stored in a per-network directory. We need to create it
-        # in order to write the Noise private key
-        datadir_with_network = os.path.join(datadir, "regtest")
-        os.makedirs(datadir_with_network, exist_ok=True)
-
-        bin = os.path.join(
-            os.path.dirname(__file__), "..", "..", "target/debug/revaultd"
-        )
-        self.conf_file = os.path.join(datadir, "config.toml")
-        self.cmd_line = [bin, "--conf", f"{self.conf_file}"]
-        socket_path = os.path.join(datadir_with_network, "revaultd_rpc")
-        self.rpc = UnixDomainSocketRpc(socket_path)
-
-        noise_secret_file = os.path.join(datadir_with_network, "noise_secret")
-        with open(noise_secret_file, "wb") as f:
-            f.write(noise_priv)
-
-        bitcoind_cookie = os.path.join(bitcoind.bitcoin_dir, "regtest", ".cookie")
-        with open(self.conf_file, "w") as f:
-            f.write(f"unvault_csv = {csv}\n")
-            f.write(f"data_dir = '{datadir}'\n")
-            f.write(f"daemon = false\n")
-            f.write(f"log_level = '{LOG_LEVEL}'\n")
-
-            f.write(f'coordinator_host = "127.0.0.1:{coordinator_port}"\n')
-            f.write(f'coordinator_noise_key = "{coordinator_noise_key}"\n')
-            f.write(f"coordinator_poll_seconds = 2\n")
-
-            f.write("stakeholders_xpubs = [")
-            for stk in stks:
-                f.write(f'"{stk.get_xpub()}", ')
-            f.write("]\n")
-
-            f.write("managers_xpubs = [")
-            for man in mans:
-                f.write(f'"{man.get_xpub()}", ')
-            f.write("]\n")
-
-            f.write("cosigners_keys = [")
-            for cosig in cosigs:
-                f.write(f'"{cosig.get_static_key().hex()}", ')
-            f.write("]\n")
-
-            f.write(f"[bitcoind_config]\n")
-            f.write(f'network = "regtest"\n')
-            f.write(f"cookie_path = '{bitcoind_cookie}'\n")
-            f.write(f"addr = '127.0.0.1:{bitcoind.rpcport}'\n")
-            f.write(f"poll_interval_secs = 3\n")
-
-            if stk_config is not None:
-                f.write("[stakeholder_config]\n")
-                self.stk_keychain = stk_config["keychain"]
-                f.write(f'xpub = "{self.stk_keychain.get_xpub()}"\n')
-                f.write("watchtowers = [")
-                for wt in stk_config["watchtowers"]:
-                    f.write(
-                        f"{{ \"host\" = \"{wt['host']}\", \"noise_key\" = "
-                        f"\"{wt['noise_key']}\" }}, "
-                    )
-                f.write("]\n")
-                # FIXME: eventually use a real one here
-                f.write(
-                    "emergency_address = "
-                    '"bcrt1qewc2348370pgw8kjz8gy09z8xyh0d9fxde6nzamd3txc9gkmjqmq8m4cdq"\n'
-                )
-
-            if man_config is not None:
-                f.write("[manager_config]\n")
-                self.man_keychain = man_config["keychain"]
-                f.write(f'xpub = "{self.man_keychain.get_xpub()}"\n')
-                f.write("cosigners = [")
-                for wt in man_config["cosigners"]:
-                    f.write(
-                        f"{{ \"host\" = \"{wt['host']}\", \"noise_key\" = "
-                        f"\"{wt['noise_key']}\" }}, "
-                    )
-                f.write("]\n")
-
-    def wait_for_deposits(self, outpoints):
-        """
-        Polls listvaults until we acknowledge the confirmed vaults at {outpoints}
-        """
-        assert isinstance(outpoints, list)
-        wait_for(
-            lambda: len(self.rpc.listvaults(["funded"], outpoints)["vaults"])
-            == len(outpoints)
-        )
-
-    def wait_for_secured_vaults(self, outpoints):
-        """
-        Polls listvaults until we acknowledge the 'secured' :tm: vaults at {outpoints}
-        """
-        assert isinstance(outpoints, list)
-        wait_for(
-            lambda: len(self.rpc.listvaults(["secured"], outpoints)["vaults"])
-            == len(outpoints)
-        )
-
-    def wait_for_active_vaults(self, outpoints):
-        """
-        Polls listvaults until we acknowledge the active vaults at {outpoints}
-        """
-        assert isinstance(outpoints, list)
-        wait_for(
-            lambda: len(self.rpc.listvaults(["active"], outpoints)["vaults"])
-            == len(outpoints)
-        )
-
-    def start(self):
-        TailableProc.start(self)
-        self.wait_for_logs(
-            [
-                "revaultd started on network regtest",
-                "bitcoind now synced",
-                "JSONRPC server started",
-                "Signature fetcher thread started",
-            ]
-        )
-
-    def stop(self):
-        try:
-            self.rpc.stop()
-            self.wait_for_logs(
-                [
-                    "Stopping revaultd.",
-                    "Bitcoind received shutdown.",
-                    "Signature fetcher thread received shutdown.",
-                ]
-            )
-            return 0
-        except Exception as e:
-            logging.error(f"{self.prefix} : error when calling stop: '{e}'")
-            return TailableProc.stop(self)
-
-    def cleanup(self):
-        try:
-            self.stop()
-        except Exception:
-            self.proc.kill()
-
-
-class ManagerRevaultd(Revaultd):
-    def __init__(
-        self,
-        datadir,
-        stks,
-        cosigs,
-        mans,
-        csv,
-        noise_priv,
-        coordinator_noise_key,
-        coordinator_port,
-        bitcoind,
-        man_config,
-    ):
-        """The wallet daemon for a manager.
-        Needs to know all xpubs, and needs to be able to connect to the
-        coordinator and the cosigners.
-        """
-        super(ManagerRevaultd, self).__init__(
-            datadir,
-            stks,
-            cosigs,
-            mans,
-            csv,
-            noise_priv,
-            coordinator_noise_key,
-            coordinator_port,
-            bitcoind,
-            man_config=man_config,
-        )
-        assert self.man_keychain is not None
-
-
-class StakeholderRevaultd(Revaultd):
-    def __init__(
-        self,
-        datadir,
-        stks,
-        cosigs,
-        mans,
-        csv,
-        noise_priv,
-        coordinator_noise_key,
-        coordinator_port,
-        bitcoind,
-        stk_config,
-    ):
-        """The wallet daemon for a stakeholder.
-        Needs to know all xpubs, and needs to be able to connect to the
-        coordinator and its watchtower(s).
-        """
-        super(StakeholderRevaultd, self).__init__(
-            datadir,
-            stks,
-            cosigs,
-            mans,
-            csv,
-            noise_priv,
-            coordinator_noise_key,
-            coordinator_port,
-            bitcoind,
-            stk_config=stk_config,
-        )
-        assert self.stk_keychain is not None
-
-
-class Coordinatord(TailableProc):
-    def __init__(
-        self,
-        datadir,
-        noise_priv,
-        managers_keys,
-        stakeholders_keys,
-        watchtowers_keys,
-        listen_port,
-        postgres_user,
-        postgres_pass,
-        postgres_host="localhost",
-    ):
-        # FIXME: reduce DEBUG log load
-        TailableProc.__init__(self, datadir, verbose=VERBOSE)
-        bin = os.path.join(
-            os.path.dirname(__file__),
-            "..",
-            "servers",
-            "coordinatord",
-            "target/debug/revault_coordinatord",
-        )
-        self.conf_file = os.path.join(datadir, "config.toml")
-        self.cmd_line = [bin, "--conf", f"{self.conf_file}"]
-        self.prefix = "coordinatord"
-
-        self.postgres_user = postgres_user
-        self.postgres_pass = postgres_pass
-        self.postgres_host = postgres_host
-        # Use the directory fixture uid
-        uid = os.path.basename(os.path.dirname(datadir))
-        self.db_name = f"revault_coordinatord_{uid}"
-        self.postgres_exec(f"CREATE DATABASE {self.db_name} OWNER {postgres_user}")
-
-        noise_secret_file = os.path.join(datadir, "noise_secret")
-        with open(noise_secret_file, "wb") as f:
-            f.write(noise_priv)
-
-        with open(self.conf_file, "w") as f:
-            f.write("daemon = false\n")
-            f.write(f'data_dir = "{datadir}"\n')
-            f.write(f'log_level = "{LOG_LEVEL}"\n')
-
-            uri = (
-                f"postgresql://{postgres_user}:{postgres_pass}"
-                f"@{postgres_host}/{self.db_name}"
-            )
-            f.write(f'postgres_uri = "{uri}"\n')
-
-            f.write("managers = [")
-            for k in managers_keys:
-                f.write(f'"{k.hex()}", ')
-            f.write("]\n")
-
-            f.write("stakeholders = [")
-            for k in stakeholders_keys:
-                f.write(f'"{k.hex()}", ')
-            f.write("]\n")
-
-            f.write("watchtowers = [")
-            for k in watchtowers_keys:
-                f.write(f'"{k.hex()}", ')
-            f.write("]\n")
-
-            f.write(f'listen = "127.0.0.1:{listen_port}"')
-
-    def postgres_exec(self, sql):
-        conn = psycopg2.connect(
-            f"dbname=postgres host={self.postgres_host} "
-            f"user={self.postgres_user} "
-            f"password={self.postgres_pass}"
-        )
-        conn.autocommit = True
-        conn.cursor().execute(sql)
-        conn.close()
-
-    def start(self):
-        TailableProc.start(self)
-        self.wait_for_logs(["Started revault_coordinatord"])
-
-    def cleanup(self):
-        try:
-            self.stop()
-        except Exception:
-            self.proc.kill()
-        self.postgres_exec(f"DROP DATABASE {self.db_name}")
-
-
-class RevaultNetwork:
-    # FIXME: we use a single bitcoind for all the wallets because it's much
-    # more efficient. Eventually, we may have to test with separate ones.
-    def __init__(
-        self,
-        root_dir,
-        bitcoind,
-        postgres_user,
-        postgres_pass,
-        postgres_host="localhost",
-    ):
-        self.root_dir = root_dir
-        self.bitcoind = bitcoind
-        self.daemons = []
-
-        self.postgres_user = postgres_user
-        self.postgres_pass = postgres_pass
-        self.postgres_host = postgres_host
-        self.coordinator_port = reserve()
-
-        self.stk_wallets = []
-        self.man_wallets = []
-
-    def deploy(self, n_stakeholders, n_managers, csv=None):
-        """
-        Deploy a revault setup with {n_stakeholders} stakeholders, {n_managers}
-        managers and optionally fund it with {funding} sats.
-        """
-        (stks, cosigs, mans) = get_participants(n_stakeholders, n_managers)
-        if csv is None:
-            # Not more than 6 months
-            csv = random.randint(1, 26784)
-
-        # The Noise keys are interdependant, so generate everything in advance
-        # to avoid roundtrips
-        coordinator_noisepriv = os.urandom(32)
-        coordinator_noisepub = bytes(
-            Curve25519Private(coordinator_noisepriv).public_key
-        )
-        (stk_noiseprivs, stk_noisepubs) = ([], [])
-        (wt_noiseprivs, wt_noisepubs) = ([], [])
-        for i in range(len(stks)):
-            stk_noiseprivs.append(os.urandom(32))
-            stk_noisepubs.append(bytes(Curve25519Private(stk_noiseprivs[i]).public_key))
-            # Unused yet
-            wt_noiseprivs.append(os.urandom(32))
-            wt_noisepubs.append(bytes(Curve25519Private(wt_noiseprivs[i]).public_key))
-        (man_noiseprivs, man_noisepubs) = ([], [])
-        for i in range(len(mans)):
-            man_noiseprivs.append(os.urandom(32))
-            man_noisepubs.append(bytes(Curve25519Private(man_noiseprivs[i]).public_key))
-        logging.debug(
-            f"Using Noise pubkeys:\n- Stakeholders: {stk_noisepubs}"
-            f"\n- Managers: {man_noisepubs}\n- Watchtowers:"
-            f"{wt_noisepubs}\n"
-        )
-
-        # Spin up the "Sync Server"
-        coord_datadir = os.path.join(self.root_dir, "coordinatord")
-        os.makedirs(coord_datadir, exist_ok=True)
-        coordinatord = Coordinatord(
-            coord_datadir,
-            coordinator_noisepriv,
-            man_noisepubs,
-            stk_noisepubs,
-            wt_noisepubs,
-            self.coordinator_port,
-            self.postgres_user,
-            self.postgres_pass,
-            self.postgres_host,
-        )
-        coordinatord.start()
-        self.daemons.append(coordinatord)
-
-        # Spin up the stakeholders wallets
-        for i in range(len(stks)):
-            datadir = os.path.join(self.root_dir, f"revaultd-stk-{i}")
-            os.makedirs(datadir, exist_ok=True)
-
-            stk_config = {
-                "keychain": stks[i],
-                # FIXME: Eventually use real ones
-                "watchtowers": [
-                    {
-                        "host": "127.0.0.1:1",
-                        "noise_key": "03c3fee141e97ed33a50875a092179684c1145"
-                        "5cc6f49a9bddaacf93cd77def697",
-                    }
-                ],
-            }
-            daemon = StakeholderRevaultd(
-                datadir,
-                stks,
-                cosigs,
-                mans,
-                csv,
-                stk_noiseprivs[i],
-                coordinator_noisepub.hex(),
-                self.coordinator_port,
-                self.bitcoind,
-                stk_config,
-            )
-            daemon.start()
-            self.stk_wallets.append(daemon)
-
-        # Spin up the managers wallets
-        for i in range(len(mans)):
-            datadir = os.path.join(self.root_dir, f"revaultd-man-{i}")
-            os.makedirs(datadir, exist_ok=True)
-
-            man_config = {
-                "keychain": mans[i],
-                # FIXME: Eventually use real ones
-                "cosigners": [
-                    {
-                        "host": "127.0.0.1:1",
-                        "noise_key": "03c3fee141e97ed33a50875a092179684c1145"
-                        "5cc6f49a9bddaacf93cd77def697",
-                    }
-                ],
-            }
-            daemon = ManagerRevaultd(
-                datadir,
-                stks,
-                cosigs,
-                mans,
-                csv,
-                man_noiseprivs[i],
-                coordinator_noisepub.hex(),
-                self.coordinator_port,
-                self.bitcoind,
-                man_config,
-            )
-            daemon.start()
-            self.man_wallets.append(daemon)
-
-        self.daemons += self.stk_wallets + self.man_wallets
-        # FIXME: we should not return them, they should access our members
-        return (self.stk_wallets, self.man_wallets)
-
-    def get_vault(self, address):
-        """Get a vault entry by outpoint or by address"""
-        for v in self.man_wallets[0].rpc.listvaults()["vaults"]:
-            if v["address"] == address:
-                return v
-
-    def fund(self, amount=None):
-        """Deposit coins into the architectures, by paying to the deposit
-        descriptor and getting the tx 6 blocks confirmations."""
-        assert len(self.man_wallets) > 0, "You must have deploy()ed first"
-
-        if amount is None:
-            amount = 49.9999
-
-        addr = self.man_wallets[0].rpc.getdepositaddress()["address"]
-        txid = self.bitcoind.rpc.sendtoaddress(addr, amount)
-        self.bitcoind.generate_block(6, wait_for_mempool=txid)
-        wait_for(lambda: self.get_vault(addr) is not None)
-
-        return self.get_vault(addr)
-
-    def secure_vault(self, vault):
-        """Make all stakeholders share signatures for all revocation txs"""
-        deposit = f"{vault['txid']}:{vault['vout']}"
-        for stk in self.stk_wallets:
-            stk.wait_for_deposits([deposit])
-            psbts = stk.rpc.getrevocationtxs(deposit)
-            cancel_psbt = stk.stk_keychain.sign_revocation_psbt(
-                psbts["cancel_tx"], vault["derivation_index"]
-            )
-            emer_psbt = stk.stk_keychain.sign_revocation_psbt(
-                psbts["emergency_tx"], vault["derivation_index"]
-            )
-            unemer_psbt = stk.stk_keychain.sign_revocation_psbt(
-                psbts["emergency_unvault_tx"], vault["derivation_index"]
-            )
-            stk.rpc.revocationtxs(deposit, cancel_psbt, emer_psbt, unemer_psbt)
-        for w in self.stk_wallets + self.man_wallets:
-            w.wait_for_secured_vaults([deposit])
-
-    def activate_vault(self, vault):
-        """Make all stakeholders share signatures for the unvault tx"""
-        deposit = f"{vault['txid']}:{vault['vout']}"
-        for stk in self.stk_wallets:
-            stk.wait_for_secured_vaults([deposit])
-            unvault_psbt = stk.rpc.getunvaulttx(deposit)["unvault_tx"]
-            unvault_psbt = stk.stk_keychain.sign_unvault_psbt(
-                unvault_psbt, vault["derivation_index"]
-            )
-            stk.rpc.unvaulttx(deposit, unvault_psbt)
-        for w in self.stk_wallets + self.man_wallets:
-            w.wait_for_active_vaults([deposit])
-
-    def stop_wallets(self):
-        for w in self.stk_wallets + self.man_wallets:
-            assert w.stop() == 0
-
-    def start_wallets(self):
-        for w in self.stk_wallets + self.man_wallets:
-            w.start()
-
-    def cleanup(self):
-        for n in self.daemons:
-            n.cleanup()
